Add SQLite support for saveing wlans adapters.
