+++ webui.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+cd /home/pi/drone-employee-connect && . venv/bin/activate && PORT=80 python serve.py
